Changing number types to int

+++ backend/calculate.py
@@ -1,15 +1,15 @@
 # Description: This file contains the functions to perform basic arithmetic operations.
 
-def add_numbers(a: float, b: float) -> float:
+def add_numbers(a: int, b: int) -> int:
     return a + b
 
-def subtract_numbers(a: float, b: float) -> float:
-    return a - b
+def subtract_numbers(a: int, b: int) -> int:
+    return int(a - b)
 
-def multiply_numbers(a: float, b: float) -> float:
-    return a * b
+def multiply_numbers(a: int, b: int) -> int:
+    return int(a * b)
 
-def divide_numbers(a: float, b: float) -> float:
+def divide_numbers(a: int, b: int) -> int:
     if b == 0:
         raise ValueError("Division by zero is not allowed")
-    return a / b
+    return int(a / b)
